requirements.txt
# Topic Modeling - arXiv cs.AI with BERTopic
# CS5660 Final Project

# Core BERTopic dependencies
bertopic>=0.16.0
sentence-transformers>=2.2.2
umap-learn>=0.5.4
hdbscan>=0.8.33

# Machine learning & numerics
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.11.0

# Data handling
pandas>=2.0.0

# arXiv API
arxiv>=2.1.0

# Text processing
spacy>=3.7.0
nltk>=3.8.0

# Evaluation metrics
gensim>=4.3.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Progress bars
tqdm>=4.66.0

# Jupyter/Colab support
ipywidgets>=8.1.0

# Optional: for faster similarity search
# faiss-cpu>=1.7.4  # Uncomment if needed
